refactor : domain 및 service 패키지 도입
