+++ src/main/java/wooteco/subway/line/dao/LineDao.java
@@ -9,7 +9,7 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import wooteco.subway.line.Line;
+import wooteco.subway.line.domain.Line;
 
 @Repository
 public class LineDao {
