+++ src/main/java/wooteco/subway/line/domain/Name.java
@@ -1,19 +1,19 @@
-package wooteco.subway.line;
+package wooteco.subway.line.domain;
 
 import java.util.Objects;
-import wooteco.subway.exception.line.NullColorException;
+import wooteco.subway.exception.NullNameException;
 
-public class Color {
+public class Name {
     private final String value;
 
-    public Color(final String value) {
+    public Name(final String value) {
         this.value = value;
         validateNull(this.value);
     }
 
     private void validateNull(final String value) {
         if (Objects.isNull(value)) {
-            throw new NullColorException();
+            throw new NullNameException();
         }
     }
