+++ src/main/java/wooteco/subway/station/domain/Station.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station;
+package wooteco.subway.station.domain;
 
 import java.util.Objects;
 import wooteco.subway.domain.Id;
